requirements.txt
Django==4.2.1
django-cors-headers==4.5.0
django-js-asset==2.2.0
djangorestframework==3.15.2
googletrans==4.0.0rc1
gTTS==2.5.3
